All tests files have .tests.cpp suffix. Also moved tests out of TestMain.cpp and moved up a level
