+++ projects/SelfTest/IntrospectiveTests/String.tests.cpp
@@ -165,3 +165,36 @@ TEST_CASE( "StringRef", "[Strings]" ) {
         }
     }
 }
+
+TEST_CASE( "replaceInPlace" ) {
+    std::string letters = "abcdefcg";
+    SECTION( "replace single char" ) {
+        CHECK( Catch::replaceInPlace( letters, "b", "z" ) );
+        CHECK( letters == "azcdefcg" );
+    }
+    SECTION( "replace two chars" ) {
+        CHECK( Catch::replaceInPlace( letters, "c", "z" ) );
+        CHECK( letters == "abzdefzg" );
+    }
+    SECTION( "replace first char" ) {
+        CHECK( Catch::replaceInPlace( letters, "a", "z" ) );
+        CHECK( letters == "zbcdefcg" );
+    }
+    SECTION( "replace last char" ) {
+        CHECK( Catch::replaceInPlace( letters, "g", "z" ) );
+        CHECK( letters == "abcdefcz" );
+    }
+    SECTION( "replace all chars" ) {
+        CHECK( Catch::replaceInPlace( letters, letters, "replaced" ) );
+        CHECK( letters == "replaced" );
+    }
+    SECTION( "replace no chars" ) {
+        CHECK_FALSE( Catch::replaceInPlace( letters, "x", "z" ) );
+        CHECK( letters == letters );
+    }
+    SECTION( "escape '" ) {
+        std::string s = "didn't";
+        CHECK( Catch::replaceInPlace( s, "'", "|'" ) );
+        CHECK( s == "didn|'t" );
+    }
+}
